Move styled macro to src/react
